Fix(app-logic): fix app logic
- Fix application logic to always generate 10 digits
- save files as csv
- add bullet numbers to generated numbers

[Fixes #167416796]
