+++ src/Helpers/sortAscending.js
@@ -1,0 +1,3 @@
+export const sortAscending = (generatedNumbers) => {
+  return generatedNumbers.sort();
+};
