+++ src/Helpers/sortDescending.js
@@ -1,0 +1,3 @@
+export const sortDescending = (generatedNumbers) => {
+  return generatedNumbers.sort((a, b) => parseInt(b) - parseInt(a));
+};
